Switched order of pairs on meanBranchDistance output

+++ src/phylo.hpp
@@ -39,7 +39,7 @@ class Phylogeny {
   ///Counts the number of species which are alive at a given point in time
   int numAlive(double t) const;
   ///Calculate the mean branch distance for the phylogeny
-  std::vector< std::pair<unsigned int,double> > meanBranchDistance(double t) const;
+  std::vector< std::pair<double, unsigned int> > meanBranchDistance(double t) const;
 };
 
 #endif
